Refactor forms into modular files with shared styling mixin
