+++ inventory/forms/bulk_forms.py
@@ -1,0 +1,9 @@
+from django import forms
+
+
+class BulkUploadForm(forms.Form):
+    file = forms.FileField()
+
+
+class BulkDeleteForm(forms.Form):
+    file = forms.FileField()
