+++ inventory/forms/stock_forms.py
@@ -1,0 +1,60 @@
+from django import forms
+
+from ..models import StockTransaction
+from .base import StyledFormMixin
+
+
+class StockReceivingForm(StyledFormMixin, forms.ModelForm):
+    class Meta:
+        model = StockTransaction
+        fields = ["item", "quantity_change", "user_id", "related_po_id", "notes"]
+        labels = {
+            "quantity_change": "Quantity",
+            "related_po_id": "PO ID",
+        }
+
+    def save(self, commit: bool = True):
+        obj = super().save(commit=False)
+        obj.transaction_type = "RECEIVING"
+        if commit:
+            obj.save()
+        return obj
+
+
+class StockAdjustmentForm(StyledFormMixin, forms.ModelForm):
+    class Meta:
+        model = StockTransaction
+        fields = ["item", "quantity_change", "user_id", "notes"]
+        labels = {"quantity_change": "Quantity Change"}
+
+    def save(self, commit: bool = True):
+        obj = super().save(commit=False)
+        obj.transaction_type = "ADJUSTMENT"
+        if commit:
+            obj.save()
+        return obj
+
+
+class StockWastageForm(StyledFormMixin, forms.ModelForm):
+    class Meta:
+        model = StockTransaction
+        fields = ["item", "quantity_change", "user_id", "notes"]
+        labels = {"quantity_change": "Quantity"}
+
+    def clean_quantity_change(self):
+        qty = self.cleaned_data.get("quantity_change")
+        if qty is None or qty <= 0:
+            raise forms.ValidationError("Quantity must be positive")
+        return qty
+
+    def save(self, commit: bool = True):
+        obj = super().save(commit=False)
+        obj.transaction_type = "WASTAGE"
+        obj.quantity_change = -abs(obj.quantity_change or 0)
+        if commit:
+            obj.save()
+        return obj
+
+
+class StockBulkUploadForm(forms.Form):
+    file = forms.FileField()
