+++ inventory/views/recipes.py
@@ -2,7 +2,7 @@ from django.contrib import messages
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.generic import TemplateView
 
-from ..forms import RecipeForm, RecipeComponentFormSet
+from ..forms.recipe_forms import RecipeForm, RecipeComponentFormSet
 from ..models import Recipe
